# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AuthData:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  dto.AuthSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.AuthData'
      message:
        example: login success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.CommentData:
    properties:
      author_id:
        example: 1
        type: integer
      content:
        example: Task nay can lam gap
        type: string
      created_at:
        type: string
      id:
        example: 1
        type: integer
      task_id:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  dto.CommentListPayload:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CommentData'
        type: array
      pagination:
        $ref: '#/definitions/dto.SwaggerPaginationResponse'
    type: object
  dto.CommentListSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.CommentListPayload'
      message:
        example: get comments success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.CommentSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.CommentData'
      message:
        example: create comment success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.CreateCommentRequest:
    properties:
      content:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: request failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.MeData:
    properties:
      created_at:
        type: string
      email:
        example: leanh@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Le Anh
        type: string
      role:
        example: member
        type: string
    type: object
  dto.MeSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.MeData'
      message:
        example: get profile success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.SimpleSuccessResponse:
    properties:
      message:
        example: delete success
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.SwaggerPaginationResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 25
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  dto.UpdateCommentRequest:
    properties:
      content:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: RESTful API cho hệ thống Task Management
  title: Task Management API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập và lấy access token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Lấy thông tin user hiện tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Đăng ký tài khoản mới
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register account
      tags:
      - Auth
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SimpleSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Cập nhật comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - Comments
...
  /tasks/{id}/comments:
    get:
      consumes:
      - application/json
      description: Lấy danh sách comment theo task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CommentListSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List comments by task
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Tạo comment cho task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CommentSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - Comments
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"